Move Theano-dependent code in sampler package
The class ReplayPool found in module utils from garage.sampler, was
extracted from the module and moved to garage.theano.
The rest of the modules were not modified.
